Moved json serialization to a separate class. Will repeat for other customizables. Also, work on making it easier to work with customizables
